Preserve hostnames in the SRV tuple
When the SRV records refer to TLS-protected services, always knowing
the name of the service (as present in the SRV RR) is important so as to
properly verify the connection.
This commit modifies the tuple structure to add a hostname field (in
addition to the host field) which contains the original hostname from
the RR.

Where the name to IP mapping fails (perhaps because the name isn't
resolvable), the SRV tuple will contain the same value in both the host
and hostname fields).

The hostname field is added to the end of the tuple to minimise backward
compatibility issues.